backend/requirements.txt
fastapi>=0.109.2
uvicorn>=0.27.1
gunicorn>=21.2.0
pydantic>=2.6.1
pydantic-settings>=2.2.1
psycopg2-binary>=2.9.9
redis>=5.0.1
requests>=2.31.0
httpx>=0.27.0
pytest>=8.0.2
pytest-asyncio>=0.23.5
